requirements.txt
torch==2.7.1
albumentations==2.0.8
matplotlib==3.10.3
numpy==2.3.0
pillow==11.2.1
scikit-learn==1.7.0
pandas==2.3.0
tqdm==4.67.1
seaborn==0.13.2
torchvision==0.22.1
faiss-cpu==1.11.0
open_clip_torch==2.32.0
jupyterlab==4.4.3
sentence_transformers==4.1.0
datasets==3.6.0
hf-xet==1.1.3
sacremoses==0.1.1
tensorboard==2.19.0
timm==1.0.15